+++ internal/xds/types/resourceversiontable.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/envoyproxy/go-control-plane/pkg/cache/types"
 	resourcev3 "github.com/envoyproxy/go-control-plane/pkg/resource/v3"
-	protobuf "google.golang.org/protobuf/proto"
 
 	"github.com/envoyproxy/gateway/internal/ir"
 	"github.com/envoyproxy/gateway/internal/utils/proto"
@@ -26,45 +25,6 @@ type EnvoyPatchPolicyStatuses []*ir.EnvoyPatchPolicyStatus
 type ResourceVersionTable struct {
 	XdsResources
 	EnvoyPatchPolicyStatuses
-}
-
-// DeepCopyInto copies the contents into the output object
-// This was generated by controller-gen, moved from
-// zz_generated.deepcopy.go and updated to use proto.Clone
-// to deep copy the proto.Message
-func (t *ResourceVersionTable) DeepCopyInto(out *ResourceVersionTable) {
-	*out = *t
-	if t.XdsResources != nil {
-		in, out := &t.XdsResources, &out.XdsResources
-		*out = make(map[resourcev3.Type][]types.Resource, len(*in))
-		for key, val := range *in {
-			var outVal []types.Resource
-			if val == nil {
-				(*out)[key] = nil
-			} else {
-				// Snippet was generated by controller-gen
-				// G601: Implicit memory aliasing in for loop.
-				in, out := &val, &outVal //nolint:gosec,scopelint
-				*out = make([]types.Resource, len(*in))
-				for i := range *in {
-					(*out)[i] = protobuf.Clone((*in)[i])
-				}
-			}
-			(*out)[key] = outVal
-		}
-	}
-}
-
-// DeepCopy generates a deep copy of the ResourceVersionTable object.
-// This was generated by controller-gen and moved over from
-// zz_generated.deepcopy.go to this file.
-func (t *ResourceVersionTable) DeepCopy() *ResourceVersionTable {
-	if t == nil {
-		return nil
-	}
-	out := new(ResourceVersionTable)
-	t.DeepCopyInto(out)
-	return out
 }
 
 // GetXdsResources retrieves the translated xds resources saved in the translator context.
